requirements.txt
pyPDF2
sentence-transformers
faiss-cpu
numpy
pandas
streamlit
rapidfuzz
